subscriptions: handle missing contact list
Some Lync installations do not react to subsription to roaming contacts.
They require that we indicate support for UCS.

When we add the necessary header to the subscription the server can
reject it with 488. When this happens then lets assume that we are
talking to a Lync 2013 installation that supports UCS.

Triggered by the discussion in

  https://sourceforge.net/p/sipe/discussion/688534/thread/cf174c35/