purple: full debug when unsafe flag is enabled
If supported by the libpurple API we now use purple_debug_is_unsafe()
as the flag to indicate full debugging. This can be enabled by calling
pidgin like this

   PURPLE_UNSAFE_DEBUG=1 pidgin ...

This means that message contents, which can contain secret information,
will only be dumped to the debug log if enabled explicitly by the user.