lync: fix crash for delayed completion of request
Observed in my test setup:

 - SSL certificate for SERVER is self-signed
 - SSL certificate for SERVER not accepted by user yet
 - request A for http://SERVER initiated
 - request B for https://SERVER initiated
 - UI pops up certificate accept/reject dialog for SERVER
 - request A completes, causing requests A & B to be freed
 - user answers dialog
 - backend is called with pointer to deleted transport -> *BOOM*

Fix this by marking parallel requests as invalid by setting the callback
to NULL when one of the request completes. The invalid requests will be
freed after backend is finished with them, e.g. the connection is
aborted when the user presses cancel on the dialog.