lync: assign unique ID to parallel requests
Use the address of the first autodiscover request data structure as
unique ID and assign it to all parallel requests. This makes sure that
one completion of one request we will be able to identify the other
request from the list.

If the ID is NULL then the request is no longer active.