appshare: refactor client interface
The RDP client data structure is an implementation detail that has now
business to appear in the sipe-appshare interface.  Factor it out to
sipe-appshare-client.h and also obsolete the header files per client
implementation in the same go.

This also takes care of the build failure observed when trying to
cross-compile for the Windows platform.