appshare: fix more coverity warnings
Re-ran analysis with a derived model for glib2. It seems that Coverity
is convinced that g_socket_get_fd() can return an invalid file
descriptor for a valid socket. Refactor the code to add an additional
error check for the file descriptor.

Found also one missing g_object_unref() when going through the code.