media: fix no relay with purple 2.x & GStreamer 1.0
Starting with version 2.10.12 it is possible to compile libpurple 2.x
with GStreamer 1.0 and Farstream 0.2, which invalidates our check
deciding whether to use GValueArray or GPtrArray as a data type of
"relay-info" property.

Moreover, because Sipe V&V has since 1.22.0 gained a build dependency
on Farstream 0.2, we may as well do away with the legacy code and use
GPtrArray everywhere now.