purple: avoid rare SSL read deadlock
I observed a rare case of a read deadlock for SSL connections. F.ex.
with an input message of 12287 bytes (= 4095 + 2 * 4096):

- setup SSL connection with a buffer of size 0
- enter transport_input_common() and extend buffer to 4096 bytes
- try to read 4095 bytes (leaves space for the string terminator)
- purple_ssl_read() reads 4095 bytes
- go around the loop and extend buffer to 8192 bytes
- try to read 4096 bytes
- purple_ssl_read() reads 4096 bytes
- go around the loop and extend buffer to 12288 bytes
- try to read 4096 bytes
- purple_ssl_read() reads 4096 bytes
- go around the loop and extend buffer to 16384 bytes
- try to read 4096 bytes
- purple_ssl_read() returns len < 0 and sets EAGAIN
- transport_input_common() exits to try again later...
- transport_input_common() is never called again -> deadlock

It seems that if purple_ssl_read() returns the exact number of bytes
requested *AND* if those are also the last pending bytes then the next
call does not return the expected length of 0.

Add special case to EAGAIN handling to avoid the deadlock.