digest: add support for OpenSSL 1.1.0
It turns out that Debian 9.0 has switched to OpenSSL 1.1.0 which by
default deprecates a lot of older APIs. That explains all those build
problems that I can't reproduce with OpenSSL 1.0.1.

For HMAC digest code we need to check the OpenSSL version number, as
the new API functions aren't available in older versions.