media: fix multpart SDP processing of phone call on SfB
User has observed a phone gateway in SfB environment that tags the SDP
part of its initial SIP INVITE with "ms-proxy-2007fallback" although it
is the only SDP payload in the message and is in RFC 5245 format.

Since apparently ms-proxy-2007fallback label can't be relied upon,
detect the SDP protocol version from the contents of the message body.