OBS: reverse logic in .dsc creation
Make the newest .dsc, i.e. with gstreamer1.0, the default and override
it for older platforms, i.e. thosw which only support telepathy, with a
platform specific .dsc file. This approach is similar to the one in the
RPM SPEC file. It requires less maintenance, because newer platforms
should now work "out of the box" immediately when they are added to the
OBS project.