purple: add support for latest appstreamcli
A few days ago a change was pushed to AppStream upstream that adds a
validation check for URLs. Unfortunately build machines may not have
access to the Internet and the validation fails.

Add another version of the validation command that adds the --nonet
option. We now try from the newest to the oldest validation command
line.