Fix #337: Duplicate candidates in SDP
This reverts commit 89ee8e8d4fd7b90e38c53c2458853389b0cbd182.

libnice can return a candidate list with duplicates. It is currently
unknown if this is a bug in libnice or a configuration error in Skype
for Business setups.

While this is not a bug in SIPE, by converting the previous debug code
into a filter function that removes these duplicates we can harden SIPE
against such a situation and make sure that it generates a correct SDP
message.

(cherry picked from commit 50f63c7d337f57374d363018c2316449a73c5449)