media: create a failed_media for all rejected streams
The Windows Skype for Business client expects that the number of medias in
the SDP message remains the same during one call.
If we start a call with audio and video and the Windows client only accepts
a audio connection then the video stream must remain as a failed_media.
Otherwise the connection setup will not finish.