purple: fix double unref of relay info array
This code was a relic from back when media_relays used to be
a GValueArray. Assigning it to "relay-info" GParameter created a copy
of the whole array which had to be freed after use.

Farstream now uses GPtrArray, which is reference-counted and we should
unref it only once in sipe_backend_media_relays_free() in order to avoid
premature deallocation.

Fixes memory corruption and occasional crashes happening mostly after
remote application sharing has been closed.