purple: add call to purple_presence_set_idle()
One of the idiosyncrasies in the libpurple API: it is not enough to set
the buddy status to one of the idle ones. You also must make another
call to set an idle flag!

User visible change of behaviour: buddies flagged as idle are shown by
Pidgin "greyed out" like offline buddies.