buddy: update sipe_core_buddy_got_status() API
Add a new time_t parameter that provides an indication to the backend
when the buddy has entered the current idle state. This parameter is
only valid when the "activity" indicates an idle state. The parameter
may be 0 if it is unknown when the buddy entered the idle state or he
is not idle.

Pass the new parameter down to sipe_backend_buddy_set_status().