rtf: add module for RTF processing
This adds a function to extract plain text from RTF and convert it to
HTML. The implementation will use a flex-based lexer and a hand-written
parser.

Start off by adding the lexer module and integrating it to the build.
The release tarball will include the pre-generated lexer C code, i.e.
the flex tool is not required for a normal build. Update the
documentation accordingly.