purple: discard the use of the global stun server
When the user has configured a stun server in the network perferences of
pidgin, it'll be passed to the farstream backend automatically, and may
generate srflx local candidates, that fail to be nominated by SfB, when
pidgin/sipe is in controlled mode (ie when in the position a receive a
stream from a SfB peer), and behind a NAT-ted network.

ms-client-diagnostics: 21
"Call failed to establish due to a media connectivity
failure where one endpoint is of unknown type"

A supplementary condition is that the NAT does not randomize ports
allocation, so the srflx candidate remains usable during the conncheck.

Just removing the possibily to generate these srflx candidates at
gathering time, by discarding the use of a custom stun server seems to
resolve this issue.  This is not a loss a pairs combinations, because in
the typical situation, the srfx local candidate discovered by the stun
server, is also discovered by the conncheck, as a prflx local candidate
this time.  And a candidate of this type is happily accepted by SfB.