adium: add RTF module to build
Add sipe-rtf.l to the Xcode build.

The default "System Lex rule" doesn't work for us as it expects the
wrong output file name in the wrong location. Create a custom build
rule using a script that calls $(LEX), which is actually "flex", and
generates files in the derived file directory with our file name
convention.