media: use purple purple_media_set_require_encryption api
This patch uses the require-encryption property of the farstream
conference element, to block by default the stream at the srtpdec
element at creation time, until we know from sdp whether it will be
encrypted or not. The installation of the keys in the element unblocks
the stream, by processing it through the decoder, else the toggle of the
property to "false" also unblocks the stream in "pass through" mode.

This patch prevents a crypted stream to pass through the srtpdec element
in the short time frame before the crypto elements have been installed,
and to reach the rest of the pipeline, causing codec decoder errors
where decrypted stream is expected.

https://gitlab.freedesktop.org/gstreamer/gst-plugins-bad/issues/1016